Updated the device header with the new peripheral addresses.
Added bss section clearing in the startup file.

The elf views and object views are now stored in files in the build folder.

Added memory erasing and reading option to the upload_prog.py python script.

After fixing the mistake in the memory addressing, this code executes successfully with the current compiler flags.